Stronger typings for coordinates and added query for multiple coordinates

+++ server/src/models/seedData.ts
@@ -3,9 +3,24 @@ import logger from '../logger';
 import mongoose from 'mongoose';
 import {
   variantAnnotationSchema,
+  VariantAnnotationCoordinates,
   VariantAnnotationDocument,
   VariantAnnotationModel,
 } from './variantAnnotation';
+
+const generateRandomCoordinates: () => VariantAnnotationCoordinates = () => {
+  const bases = ['A', 'T', 'C', 'G'];
+  const ref = Faker.helpers.randomize(bases);
+  const alt = Faker.helpers.randomize(bases.filter(b => b !== ref));
+  const chr = Faker.helpers.randomize([...Array(24).keys()].map(x => x + 1));
+  const assembly = Faker.helpers.randomize([37, 38]);
+  return {
+    assembly: assembly,
+    alt: alt,
+    chr: chr,
+    ref: ref,
+  };
+};
 
 /**
  * @param model: Each Mongoose model is scoped to a single connection only. To use this function, a connection must be established beforehand.
@@ -16,11 +31,12 @@ const createDummyVariantAnnotations = async (model: VariantAnnotationModel) => {
   const variants = Array(5000)
     .fill(null)
     .map(() => {
+      const coordinates = generateRandomCoordinates();
       const bases = ['A', 'T', 'C', 'G'];
-      const ref = Faker.helpers.randomize(bases);
-      const alt = Faker.helpers.randomize(bases.filter(b => b !== ref));
-      const chr = Faker.helpers.randomize(['X', 'Y', ...[...Array(22).keys()].map(x => x + 1)]);
-      const assembly = Faker.helpers.randomize([37, 38]);
+      const ref = coordinates.ref;
+      const alt = coordinates.alt;
+      const chr = coordinates.chr;
+      const assembly = coordinates.assembly;
       const aaChanges = `Z[${ref}GC] > Y[${alt}GC]`;
       const cdna = Array(100)
         .fill(null)
@@ -51,6 +67,19 @@ const createDummyVariantAnnotations = async (model: VariantAnnotationModel) => {
   }
 };
 
+const queryManyCoordinates = async (model: VariantAnnotationModel) => {
+  const coordinates = Array(1000)
+    .fill(null)
+    .map(() => generateRandomCoordinates());
+  console.log(coordinates);
+  try {
+    const annotations = await model.find({ $or: coordinates }).explain('executionStats');
+    console.log(annotations);
+  } catch (err) {
+    logger.error(err);
+  }
+};
+
 const MONGO_URL = `mongodb://${process.env.MONGO_INITDB_ROOT_USERNAME}:${process.env.MONGO_INITDB_ROOT_PASSWORD}@localhost:${process.env.MONGO_INITDB_PORT}`;
 
 mongoose.connect(MONGO_URL!).then(async () => {
@@ -58,16 +87,20 @@ mongoose.connect(MONGO_URL!).then(async () => {
     'VariantAnnotation',
     variantAnnotationSchema
   );
-  await Promise.all([model.deleteMany({})]);
-  createDummyVariantAnnotations(model);
-  model.createIndexes([
-    {
-      assembly: 1,
-      alt: 1,
-      chr: 1,
-      ref: 1,
-    },
-  ]);
+  const eraseDatabaseOnSync = false;
+  if (eraseDatabaseOnSync) {
+    await Promise.all([model.deleteMany({})]);
+    createDummyVariantAnnotations(model);
+    model.createIndexes([
+      {
+        assembly: 1,
+        alt: 1,
+        chr: 1,
+        ref: 1,
+      },
+    ]);
+  }
+  queryManyCoordinates(model);
 });
 
 export default createDummyVariantAnnotations;
